refactor: computer 유틸리티 적용
