+++ docs/README.md
@@ -14,7 +14,7 @@
 
   - `lap`인스턴스에 자동차 이름 배열과, 실행횟수를 전달한다.
 
-  - `lap`의 `printStage()` 메서드를 실행횟수 만큼 실행한다.
+  - `lap`의 `printStage()`메서드를 실행횟수 만큼 실행한다.
 
   - `lap`의 `record`를 통해 우승자를 확인하고, 출력한다.
 
@@ -24,18 +24,37 @@
 
   - 각 차수별 실행 결과를 기록한다. `record`
 
-### 유틸리티 함수
+### 게임 로직 함수
 
-- `validation()`
+- `lap.printStage()`
 
-  - 자동차 이름 입력 시 검증 항목
+  - 차수별 진행 상황 출력
+
+- `printResult()`
+
+  - 최종 결과 배열에서 우승자 판단 후 출력
+
+- `validatate()`
+
+  - 자동차 이름 입력 시 검증 항목 `carName`
 
     - 입력이 최소 1개에서 최대 5자 사이로 들어오게
 
     - 최대 9대까지만 등록이 가능하도록
 
-  - 실행횟수 입력 시 검증 항목
+  - 실행횟수 입력 시 검증 항목 `lapLength`
 
     - 숫자만 입력받도록
 
     - 최대 99까지 입력이 가능하도록
+
+  ### 유틸리티 함수
+
+- `console.js`
+
+    - `Console` 유틸리티 추상화, 네이밍 변경
+
+- `constanst.js`
+
+    - 상수 메세지 전달
+
